RDKB-41327 - Implement Round Trip Time Tracker in RBUS Daemon
Reason for change: To solve the issue with XE2 pod, as it is musl gcc it is expecting
time_t as a long long int so type casting the variable to long long int with format
specifier as lld to work with musl and normal gcc
Test Procedure: Test and verified
Risks: Low

+++ rtrouted.c
@@ -182,7 +182,7 @@ rtRouted_TransactionTimingDetails(rtMessageHeader header_details)
   timestamp.tv_sec = header_details.T5 + boottime;
   rtTime_ToString(&timestamp, time_buff);
   printf("Time at which daemon received the response             : %s\n", time_buff);
-  printf("Total duration                                         : %ld seconds\n", (header_details.T5 - header_details.T1));
+  printf("Total duration                                         : %lld seconds\n", (long long int)(header_details.T5 - header_details.T1));
   printf("=======================================================================\n");
 }
 #endif
